requirements.txt
torch
torchdyn
pytorch-lightning
functorch
matplotlib
tqdm
wandb
python-box
mat73
gdown